User can now add spaces to their profile so there can be a blank column in the Excel file generated.
Fixed an issue where the OC and RC sheets couldn't be made as it tried to recreate the same one twice, thus i created CreateMasterSheet().

There is still a bug where the headerCounter in ProfileCreator.cs does not properly work. Temporarily disabled the - button so you wouldn't get of the trace.